requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
black
pytest
python-dotenv>=0.5.1
ipykernel

# data
streamlit
pandas
kedro-datasets

# cv
opencv-python==4.6.0.66 
ultralytics
torch-cpu
torchvision